Extract `TypeMapping` interface
